[REFACTOR] separate component to different file
